# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "workharbor API",
    "version": "1",
    "description": "The JSON API of `whr serve` (design \u00a79.7). Loopback only; every route needs the bearer token. Responses are wrapped in the envelope of \u00a79.2. Free text from agents, repositories and the forge (input, subject, message) is untrusted data."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/v1/openapi.json": {
      "get": {
        "operationId": "openapi",
        "summary": "This document",
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "The OpenAPI document (not wrapped in the envelope)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "An error. The HTTP status follows the exit code: 400 usage (2), 401 auth (4), 404 not found (3), 409 conflict or needs human (5, 6), 504 timeout (7), 500 other.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "schema_version",
          "ok",
          "error"
        ],
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1
          },
          "ok": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "exit_code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "exit_code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}